+++ client/src/config/eventSearchTags.ts
@@ -12,4 +12,58 @@ export const relatedObjects: IHalLinks = {
     searchTag: 'lux:eventIncludedItems',
     tab: 'objects',
   },
+  works: {
+    title: 'Works About',
+    searchTag: 'lux:eventWorksAbout',
+    tab: 'works',
+  },
+}
+
+// Used for rendering related types accordions
+export const itemAndWorkTypes: IHalLinks = {
+  itemTypes: {
+    title: 'Object Types',
+    searchTag: 'lux:eventObjectTypesUsed',
+    tab: 'objects',
+    jsonSearchTerm: 'classification',
+  },
+  itemTypesPeriods: {
+    title: 'Object Types',
+    searchTag: 'lux:eventObjectTypesAbout',
+    tab: 'objects',
+    jsonSearchTerm: 'classification',
+  },
+  workTypes: {
+    title: 'Work Types',
+    searchTag: 'lux:eventWorkTypesUsed',
+    tab: 'works',
+    jsonSearchTerm: 'classification',
+  },
+  workTypesPeriods: {
+    title: 'Work Types',
+    searchTag: 'lux:eventWorkTypesAbout',
+    tab: 'works',
+    jsonSearchTerm: 'classification',
+  },
+}
+
+// Shows related accordions
+export const relatedAccordions: IHalLinks = {
+  conceptInfluenced: {
+    title: 'Concepts Influenced',
+    searchTag: 'lux:eventConceptsInfluencedBy',
+    tab: 'concepts',
+  },
+  materials: {
+    title: 'Related Materials',
+    searchTag: 'lux:eventItemMaterials',
+    tab: 'concepts',
+    jsonSearchTerm: 'material',
+  },
+  subjects: {
+    title: 'Related Subjects',
+    searchTag: 'lux:eventConceptsOfItems',
+    tab: 'concepts',
+  },
+  ...itemAndWorkTypes,
 }