+++ client/src/features/advancedSearch/TextInput.tsx
@@ -7,6 +7,10 @@ import EntityParser from '../../lib/parse/data/EntityParser'
 import { useGetNameQuery } from '../../redux/api/ml_api'
 import { addTextValue } from '../../redux/slices/advancedSearchSlice'
 import { StyledInput } from '../../styles/features/advancedSearch/Input'
+import {
+  addHoverHelpText,
+  addSelectedHelpText,
+} from '../../redux/slices/helpTextSlice'
 // import theme from '../../styles/theme'
 
 interface IInputType {
@@ -15,6 +19,7 @@ interface IInputType {
   field: string
   stateId: string
   autoFocus?: boolean
+  scope?: string
 }
 
 /**
@@ -24,6 +29,7 @@ interface IInputType {
  * @param {string} parentScope the scope of the parent object
  * @param {string} stateId id of the current object within the advanced search state
  * @param {boolean} autoFocus optional; move keyboard focus onto the input field
+ * @param {string} scope optional; the scope of the row for updating the help text
  * @returns {JSX.Element}
  */
 const TextInput: React.FC<IInputType> = ({
@@ -32,10 +38,18 @@ const TextInput: React.FC<IInputType> = ({
   field,
   stateId,
   autoFocus,
+  scope,
 }) => {
   const dispatch = useAppDispatch()
   const handleOnChange = (userInput: string): void => {
     dispatch(addTextValue({ field, value: userInput, stateId }))
+  }
+
+  const handleOnSelect = (): void => {
+    if (scope !== undefined) {
+      dispatch(addSelectedHelpText({ value: field, scope }))
+      dispatch(addHoverHelpText({ value: field, scope }))
+    }
   }
 
   const uri = currentValue
@@ -69,6 +83,7 @@ const TextInput: React.FC<IInputType> = ({
           className="form-control advancedSearchInput"
           placeholder={label}
           onChange={(e) => handleOnChange(e.currentTarget.value)}
+          onSelect={() => handleOnSelect()}
           data-testid={`${field}-${stateId}-text-input`}
           id={id}
           disabled={displayName !== currentValue}